230403 정승원 이력서작성페이지, resumeMapper, languageDAO, VO Frontcontroller, resumeVO 추가

+++ alli/src/main/java/com/smhrd/model/LanguageVO.java
@@ -1,8 +1,10 @@
 package com.smhrd.model;
 
+import java.math.BigDecimal;
+
 public class LanguageVO {
 	
-	private int resume_num;
+	private BigDecimal resume_num;
 	private String language_name;
 	private String language_level;
 	
@@ -10,14 +12,14 @@ public class LanguageVO {
 		
 	}
 		
-	public LanguageVO(int resume_num, String language_name, String language_level) {
+	public LanguageVO(BigDecimal resume_num, String language_name, String language_level) {
 		super();
 		this.resume_num = resume_num;
 		this.language_name = language_name;
 		this.language_level = language_level;
 	}
 
-	public int getResume_num() {
+	public BigDecimal getResume_num() {
 		return resume_num;
 	}
